upgrade Kotlin and Gradle and Android Studio.

+++ floating-window-core/src/main/java/jp/kght6123/floating/window/core/FloatWindowInfo.kt
@@ -228,8 +228,8 @@ class FloatWindowInfo(
     }
     fun isOnTouchEvent(event: MotionEvent): Boolean {
         val params = getWindowLayoutParams()
-        return event.rawX in params.leftMargin..(params.leftMargin + params.width)
-                && event.rawY in params.topMargin..(params.topMargin + params.height)
+        return event.rawX as Int in params.leftMargin..(params.leftMargin + params.width)
+                && event.rawY as Int in params.topMargin..(params.topMargin + params.height)
     }
     fun updateAnchorColor(strokeMode: Stroke) {
